[otbn,dv] Picks up unaligned addresses whie executing bne and be
Following additions/ modifications are made in this commit:
1. bad_branch.py: New modes are added to pick up random unaligned
addresses
2. untaken_branch.py: Generates instructions having registers with
unequal values for beq and registers with equal values for bne.

Fixes #8086
